+++ modules/core/src/main/scala/io/bartholomews/discogs4s/entities/Pagination.scala
@@ -1,0 +1,5 @@
+package io.bartholomews.discogs4s.entities
+
+case class Pagination(page: Int, pages: Int, items: Long, per_page: Int, urls: PageUrls)
+
+case class PageUrls(first: Option[String], prev: Option[String], next: String, last: String)