+++ modules/core/src/main/scala/io/bartholomews/discogs4s/entities/UserIdentity.scala
@@ -1,0 +1,5 @@
+package io.bartholomews.discogs4s.entities
+
+import sttp.model.Uri
+
+case class UserIdentity(id: Long, username: String, resourceUrl: Uri, consumerName: String) extends DiscogsEntity