+++ modules/core/src/main/scala/io/bartholomews/discogs4s/entities/Username.scala
@@ -1,0 +1,8 @@
+package io.bartholomews.discogs4s.entities
+
+case class Username(value: String) extends AnyVal
+case class UserEmail(value: String) extends AnyVal
+case class UserRealName(value: String) extends AnyVal
+case class UserWebsite(value: String) extends AnyVal
+case class UserLocation(value: String) extends AnyVal
+case class UserProfileInfo(value: String) extends AnyVal