+++ project/CompilerOptions.scala
@@ -1,7 +1,8 @@
 object CompilerOptions {
 
-  // TODO 
+  // TODO plugin?
   lazy val compilerOptions: Seq[String] = Seq(
+    "-Ymacro-annotations", // https://github.com/circe/circe/issues/975
     "-encoding", "utf8", // Option and arguments on same line
     "-Xfatal-warnings",  // New lines for each options
     "-deprecation",
@@ -11,5 +12,4 @@ object CompilerOptions {
     "-language:existentials",
     "-language:postfixOps"
   )
-
 }