+++ src/test/scala/io/bartholomews/discogs4s/CoreWireWordSpec.scala
@@ -1,8 +1,0 @@
-package io.bartholomews.discogs4s
-
-import io.bartholomews.scalatestudo.WireWordSpec
-import org.scalatest.matchers.should.Matchers
-
-trait CoreWireWordSpec extends WireWordSpec with Matchers {
-  override val testResourcesFileRoot = "src/test/resources"
-}