+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Migrated to sttp v3
+- Multi-module projects setup (circe / play / core)
   
 ## [0.1.1] - 2021-01-02
 ### Changed